CLERN FDS - "file_server.py structured"
- FIle server should be ready for addition of motion detection.
-- Needs actual testing couldn't do it because I am on a desktop with no camera and am too lazy to alter my client end right now.
- Made some methods 'private' in tcp_server and tcp_client.

+++ Server/file_server.py
@@ -5,27 +5,42 @@ Last Modified: 3 Jul 2020
 Description: This module serves as the start script for the
              CLERN file server component of the CLERN Server
 '''
+from concurrent.futures.process import ProcessPoolExecutor
+from concurrent.futures.thread import ThreadPoolExecutor
 from Server.tcp_server import TCPServer
 from Server.archive import Archive
 import os
 import sys
 
 
-def start_server():
+def main():
     clern_server = TCPServer()
-    clern_server.connect()
+    ThreadPoolExecutor().submit(clern_server.listenLoop)
     parent_dir = os.getcwd()
     if not (os.path.exists("./archives")):
         os.mkdir("./archives")
-    try:
-        while True:
-            archive_name = clern_server.receiveFile()
-            frame_archive = Archive(archive_name)
-            os.chdir("./archives")
-            frame_archive.extract()
-            os.chdir(parent_dir)
-            os.remove("./%s" % frame_archive.file_name)
-    except Exception as err_type:
-        print("\n***TCP SERVER %s error thrown during file transfer ***" % err_type)
-        clern_server.close()
-        sys.exit()
+    while True:
+        # Actively listens for new zips
+        if len(clern_server.receivedZips) > 0:
+            try:
+                while True:
+                    archive_name = clern_server.receivedZips.pop(0)
+                    frame_archive = Archive(archive_name)
+                    os.chdir("./archives")
+                    frame_archive.extract()
+                    os.chdir(parent_dir)
+                    os.remove("./%s" % frame_archive.file_name)
+                    print(f"{archive_name} unzipped and archived")
+                    # TODO put the motion detection logic below in the ProcessPoolExecutor
+                    # so if longer than expected it doesnt get hung-up on one sequence and slow down the entire process
+                    # eg.) ProcessPoolExecutor().submit(runFile, arg1, arg2)
+                    ProcessPoolExecutor().submit()
+
+            except Exception as err_type:
+                print("\n***TCP SERVER %s error thrown during file transfer ***" % err_type)
+                clern_server.close()
+                sys.exit()
+
+
+if __name__ == '__main__':
+    main()
